Feature/tjdb table menu bar revamp

+++ frontend/src/TooljetDatabase/Table/index.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useContext, useRef, useMemo } from 'react';
+import React, { useEffect, useState, useContext, useRef } from 'react';
 import cx from 'classnames';
 import { useTable, useRowSelect } from 'react-table';
 import { isBoolean, isEmpty } from 'lodash';
@@ -21,10 +21,11 @@ import CharacterVar from '../Icons/Text.svg';
 import Boolean from '../Icons/Toggle.svg';
 import Menu from '../Icons/Menu.svg';
 import DeleteIcon from '../Table/ActionsPopover/Icons/DeleteColumn.svg';
+import TjdbTableHeader from './Header';
 
 import './styles.scss';
 
-const Table = ({ openCreateRowDrawer, openCreateColumnDrawer, collapseSidebar }) => {
+const Table = ({ collapseSidebar }) => {
   const {
     organizationId,
     columns,
@@ -42,6 +43,8 @@ const Table = ({ openCreateRowDrawer, openCreateColumnDrawer, collapseSidebar })
   const [isEditColumnDrawerOpen, setIsEditColumnDrawerOpen] = useState(false);
   const [selectedColumn, setSelectedColumn] = useState();
   const [loading, setLoading] = useState(false);
+  const [isCreateRowDrawerOpen, setIsCreateRowDrawerOpen] = useState(false);
+  const [isCreateColumnDrawerOpen, setIsCreateColumnDrawerOpen] = useState(false);
   const [editColumnHeader, setEditColumnHeader] = useState({
     hoveredColumn: null,
     clickedColumn: null,
@@ -191,22 +194,26 @@ const Table = ({ openCreateRowDrawer, openCreateColumnDrawer, collapseSidebar })
     const shouldDelete = confirm('Are you sure you want to delete the selected rows?');
     if (shouldDelete) {
       const selectedRows = Object.keys(selectedRowIds).map((key) => rows[key]);
-      const primaryKey = columns.find((column) => column.isPrimaryKey);
-      const deletionKeys = selectedRows.map((row) => {
-        return row.values[primaryKey.accessor];
-      });
+      const primaryKey = columns.find((column) => column?.constraints_type?.is_primary_key);
+      if (primaryKey) {
+        const deletionKeys = selectedRows.map((row) => {
+          return row.values[primaryKey?.accessor];
+        });
 
-      let query = `?${primaryKey.accessor}=in.(${deletionKeys.toString()})`;
+        let query = `?${primaryKey?.accessor}=in.(${deletionKeys.toString()})`;
 
-      const { error } = await tooljetDatabaseService.deleteRows(organizationId, selectedTable.id, query);
+        const { error } = await tooljetDatabaseService.deleteRows(organizationId, selectedTable.id, query);
 
-      if (error) {
-        toast.error(error?.message ?? `Error deleting rows from table "${selectedTable.table_name}"`);
-        return;
+        if (error) {
+          toast.error(error?.message ?? `Error deleting rows from table "${selectedTable.table_name}"`);
+          return;
+        }
+
+        toast.success(`Deleted ${selectedRows.length} rows from table "${selectedTable.table_name}"`);
+        fetchTableData();
+      } else {
+        toast.error('Something went wrong - Record Id is incorrect');
       }
-
-      toast.success(`Deleted ${selectedRows.length} rows from table "${selectedTable.table_name}"`);
-      fetchTableData();
     }
   };
 
@@ -345,57 +352,15 @@ const Table = ({ openCreateRowDrawer, openCreateColumnDrawer, collapseSidebar })
 
   return (
     <div>
-      <ConfirmDialog
-        title={'Delete Column'}
-        show={editColumnHeader?.deletePopupModal}
-        message={
-          'Deleting the column could affect it’s associated queries/components. Are you sure you want to continue?'
-        }
-        onConfirm={handleDeleteColumn}
-        onCancel={() => {
-          setEditColumnHeader((prevState) => ({
-            ...prevState,
-            deletePopupModal: false,
-          }));
-        }}
-        darkMode={darkMode}
-        confirmButtonType="dangerPrimary"
-        cancelButtonType="tertiary"
-        onCloseIconClick={() => {
-          setEditColumnHeader((prevState) => ({
-            ...prevState,
-            deletePopupModal: false,
-          }));
-        }}
-        confirmButtonText={'Delete Column'}
-        cancelButtonText={'Cancel'}
-        confirmIcon={<DeleteIcon />}
+      <TjdbTableHeader
+        isCreateColumnDrawerOpen={isCreateColumnDrawerOpen}
+        setIsCreateColumnDrawerOpen={setIsCreateColumnDrawerOpen}
+        isCreateRowDrawerOpen={isCreateRowDrawerOpen}
+        setIsCreateRowDrawerOpen={setIsCreateRowDrawerOpen}
+        selectedRowIds={selectedRowIds}
+        handleDeleteRow={handleDeleteRow}
+        rows={rows}
       />
-      {Object.keys(selectedRowIds).length > 0 && (
-        <div className="w-100 bg-white">
-          <button
-            onClick={handleDeleteRow}
-            type="button"
-            className="btn border-0 d-flex align-items-center delete-row-btn"
-            data-cy="delete-row-records-button"
-          >
-            <svg
-              width="13"
-              height="14"
-              viewBox="0 0 13 14"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-              data-cy="delete-row-records-icon"
-            >
-              <path
-                d="M2.97721 13.4306C2.62166 13.4306 2.31332 13.3 2.05221 13.0389C1.7911 12.7778 1.66055 12.4695 1.66055 12.1139V2.78059H1.47721C1.28832 2.78059 1.12999 2.71671 1.00221 2.58893C0.874436 2.46115 0.810547 2.30282 0.810547 2.11393C0.810547 1.92504 0.874436 1.7667 1.00221 1.63893C1.12999 1.51115 1.28832 1.44726 1.47721 1.44726H4.36055C4.36055 1.25837 4.42444 1.10004 4.55221 0.97226C4.67999 0.844483 4.83832 0.780594 5.02721 0.780594H8.39388C8.58277 0.780594 8.74388 0.847261 8.87721 0.980594C9.01055 1.11393 9.07721 1.26948 9.07721 1.44726H11.9439C12.1328 1.44726 12.2911 1.51115 12.4189 1.63893C12.5467 1.7667 12.6105 1.92504 12.6105 2.11393C12.6105 2.30282 12.5467 2.46115 12.4189 2.58893C12.2911 2.71671 12.1328 2.78059 11.9439 2.78059H11.7605V12.1139C11.7605 12.4695 11.63 12.7778 11.3689 13.0389C11.1078 13.3 10.7994 13.4306 10.4439 13.4306H2.97721ZM2.97721 2.78059V12.1139H10.4439V2.78059H2.97721ZM4.71055 10.1806C4.71055 10.3362 4.7661 10.4695 4.87721 10.5806C4.98832 10.6917 5.12166 10.7473 5.27721 10.7473C5.44388 10.7473 5.58277 10.6917 5.69388 10.5806C5.80499 10.4695 5.86055 10.3362 5.86055 10.1806V4.69726C5.86055 4.53059 5.80221 4.38893 5.68555 4.27226C5.56888 4.15559 5.43277 4.09726 5.27721 4.09726C5.11055 4.09726 4.97444 4.15559 4.86888 4.27226C4.76332 4.38893 4.71055 4.53059 4.71055 4.69726V10.1806ZM7.56055 10.1806C7.56055 10.3362 7.61888 10.4695 7.73555 10.5806C7.85221 10.6917 7.98832 10.7473 8.14388 10.7473C8.31055 10.7473 8.44944 10.6917 8.56055 10.5806C8.67166 10.4695 8.72721 10.3362 8.72721 10.1806V4.69726C8.72721 4.53059 8.66888 4.38893 8.55221 4.27226C8.43555 4.15559 8.29944 4.09726 8.14388 4.09726C7.97721 4.09726 7.83832 4.15559 7.72721 4.27226C7.6161 4.38893 7.56055 4.53059 7.56055 4.69726V10.1806ZM2.97721 2.78059V12.1139V2.78059Z"
-                fill="#FF6972"
-              />
-            </svg>
-            &nbsp; Delete records
-          </button>
-        </div>
-      )}
       <div
         style={{
           height: 'calc(100vh - 164px)', // 48px navbar + 96 for table bar +  52 px in footer
@@ -480,7 +445,7 @@ const Table = ({ openCreateRowDrawer, openCreateColumnDrawer, collapseSidebar })
                   </th>
                 ))}
                 <th
-                  onClick={() => openCreateColumnDrawer()}
+                  onClick={() => setIsCreateColumnDrawerOpen(true)}
                   className={darkMode ? 'add-icon-column-dark' : 'add-icon-column'}
                 >
                   <div className="icon-styles d-flex align-items-center justify-content-center">+</div>
@@ -597,7 +562,10 @@ const Table = ({ openCreateRowDrawer, openCreateColumnDrawer, collapseSidebar })
                 );
               })
             )}
-            <div onClick={() => openCreateRowDrawer()} className={darkMode ? 'add-icon-row-dark' : 'add-icon-row'}>
+            <div
+              onClick={() => setIsCreateRowDrawerOpen(true)}
+              className={darkMode ? 'add-icon-row-dark' : 'add-icon-row'}
+            >
               +
             </div>
           </tbody>
@@ -605,7 +573,6 @@ const Table = ({ openCreateRowDrawer, openCreateColumnDrawer, collapseSidebar })
 
         <TableFooter
           darkMode={darkMode}
-          openCreateRowDrawer={openCreateRowDrawer}
           dataLoading={loading}
           tableDataLength={tableData.length}
           collapseSidebar={collapseSidebar}
@@ -618,6 +585,32 @@ const Table = ({ openCreateRowDrawer, openCreateColumnDrawer, collapseSidebar })
           onClose={() => setIsEditColumnDrawerOpen(false)}
         />
       </Drawer>
+      <ConfirmDialog
+        title={'Delete Column'}
+        show={editColumnHeader?.deletePopupModal}
+        message={
+          'Deleting the column could affect it’s associated queries/components. Are you sure you want to continue?'
+        }
+        onConfirm={handleDeleteColumn}
+        onCancel={() => {
+          setEditColumnHeader((prevState) => ({
+            ...prevState,
+            deletePopupModal: false,
+          }));
+        }}
+        darkMode={darkMode}
+        confirmButtonType="dangerPrimary"
+        cancelButtonType="tertiary"
+        onCloseIconClick={() => {
+          setEditColumnHeader((prevState) => ({
+            ...prevState,
+            deletePopupModal: false,
+          }));
+        }}
+        confirmButtonText={'Delete Column'}
+        cancelButtonText={'Cancel'}
+        confirmIcon={<DeleteIcon />}
+      />
     </div>
   );
 };
